bool deserialization fix

+++ nd4j/nd4j-buffer/src/main/java/org/nd4j/linalg/api/buffer/BaseDataBuffer.java
@@ -1100,7 +1100,7 @@ public abstract class BaseDataBuffer implements DataBuffer {
             case BOOL:
                 for (int i = 0; i < length(); i++) {
                     try {
-                        dos.writeShort(getInt(i) == 0 ? 0 : 1);
+                        dos.writeByte(getInt(i) == 0 ? (byte) 0 : (byte) 1);
                     } catch (IOException e) {
                         e.printStackTrace();
                     }
